Make Krita's icon appear on Discover page
Because its icon isn't in the Breeze icon theme, we need to use a
locally-embedded copy of it. This necessitates swapping between a
Kirigami.Icon and an Image in the delegate depending on what kind of
thing it gets fed, since neither one of them can handle both icons form
the icon theme and also local resources.

BUG: 458397