Make the entire toolbar area draggable
To do this, we have to port away from using QQC2.ToolBar, due to a known
deficiency in it (Bug 452180). We also can't use the built-in
pageStack's header because it makes a lot of layout decisions for us and
we want a nice clean custom style here. So we use a custom
implementation of Kirigami.AbstractApplicationHeader that looks
identical to a QQC2.ToolBar, and can be used identically to is, but lets
you drag on it to move the window.

BUG: 458394