Correct GIMP's name on Discover page
"Gimp" is either a disrespectful slur for someone with physical
disabilities, or else slang for someone who participates in certain
sexual activities. The app's actual name is "GIMP," which is an
acronym standing for "GNU Image Manipulation Program". It's not the
world's best acronym, but let's at least be accurate. :)

CCBUG: 458513