Add --after-upgrade-to <plasma version> flag
This allows the app to display a post-upgrade page, so people can see
all the cool new stuff available in in the Plasma version they just
upgraded to.

Right now it's just a link to the release notes web page and a cheesy
graphic, because we don't have a data source with just a small
text-based summary of the changes. Once we have that, text from there
can be displayed right in the window itself, with the link to the
release notes providing a way to see the full announcement with
everything in it.

BUG: 459181