Re-add autostart
This is a partial revert of b1c575ce2338dde54e8a55b36904917598773e05
that brings back autostart for the primary usage mode (i.e. as an
onboarding wizard) but in a technically simpler form:
- We now use single config file key that encompass both skipping and
  completing the wizard, which are the same from the perspective of
  whether we want to autostart it next boot or not.
- We now decide whether or not to autostart on the next boot by using
  an X-KDE-autostart-condition key in the .desktop file, rather than
  conditionally deleting the autostart file

We still need to figure out a way to open the app in upgrade mode
following an upgrade to a new major Plasma version.

CCBUG: 459180