Advertise Alt+Space shortcut for KRunner, rather than Alt+F2
This shortcut is easier to trigger due to the proximity of the Alt and
spacebar keys. Additionally, the shortcut with F2 won't always work
because many laptops have the function keys bound to their secondary
functions (e.g volume up/down) by default, so when the user hits Alt+F2
on such a laptop, KRunner won't open and they'll feel frustrated.