Set "ShouldShow=false" when quitting the app using amy method
...Not just the Skip and Finish buttons. This isn't an issue in the 6.0
version of the app which uses different heuristics to determine whether
to autostart the app, but it is an isue for the 5.27 version which
simply checks for the presence of "ShouldShow=False" in the config file.

The fact that we haven't gotten absolute avalance of bug reports about
this indicates that nearly all of the people who encounter the welcome
app are either clicking on the Skip button, or going through the entire
thing and then clicking on the Finish button. So that's nice at least.

Test Plan: remove "ShouldShow=false" from ~/.config/plasma-welcomerc,
launch the app, and quit it with Ctrl+Q, Alt+F4, titlebar close button,
or Task Manager close button/menu item

BUG: 466475
FIXED-IN: 5.27.3