Make the page layouts mobile-friendly
This app was developed for a desktop use case, but people are running it
on mobile devices too. Getting it in a better shape for that use case
turns out to not be very hard--mostly a matter of replacing long row
layouts with grid or column layouts, making graphics dynamically resize,
reducing margins a bit, etc.

This commit does not adapt the *content* for mobile use cases; for that,
see https://invent.kde.org/plasma/plasma-welcome/-/issues/13.

BUG: 466572
FIXED-IN: 6.0