Make runCommand() with only one argument work as intended
Our documentation specifies this invocation, but it didn't actually work
because on the QML side, the parser errors out if you pass only one
argument, even though on the C++ side the second one is optional.

To fix this, add another method signature that only takes one argument
that just internally calls the other one with an empty string for the
second one.