Allow distros to customize elements of the welcome page
This way distros can customize it as needed without having to patch the
code, and any customizations will be integrated with the information
that KDE wants to show here.

The intro text, image, image caption, and URL can all be changed from
their default values using a specially-crafted .desktop file. This
allows for the text entries in this file to be localized by whoever
supplies the file.

As part of this, the location on disk to find customization files has
changed; now everything lives in /usr/share/plasma/plasma-welcome/. See
the README.md file for more details.

Closes #18