Don't show KAccounts page for now
In an ideal world, the KAccounts KCM would be a one-stop-shop for adding
your existing external accounts and then those accounts would be used in
all the apps and places you expect. Unfortunately that's not what
happens, and as a result user feedback about this feature has not been
very positive. For one example, see
https://discuss.kde.org/t/why-does-every-kde-pim-program-have-its-own-account-configuration-ui-why-no-kcm-for-that/6803

For now let's disable it to avoid giving users false hopes. We can
always re-enable it later in the future once KAccounts supports more
account types and integrates better with other KDE technologies.

BUG: 471380

cc @olib @teams/usability