Correct release URL for Plasma 6.0
Normally these announcements follow a predictable URL structure, but we
didn't adhere to it for the mega-release and didn't either change it
here or add a redirect on the web. The redirect is being added, but we
can also change it here to avoid pinging it all the time.

This is targeted at the 6.0 branch only; subsequent releases should use
the normal URL structure again.