Overhaul "Simple by default" and "Powerful when needed" pages
Right now the "Simple by default" page is very simple in what it shows,
but that thing is System Settings, which is decidedly *not* simple and
goes against the message that you don't need to configure anything to be
productive!

Instead, let's show a mock desktop and panel that shows information
about standard features when you hover over them. System Settings gets
moved into the "Powerful when needed" page, replacing the info about
GHNS stuff.