Use snap:// URLs rather than appstream:// ones on snap only distros
Not all the snaps in the snapstore declare proper appdata for their
package. So using only appstream:// URLs might leads to not found
packages on snap only distros like Ubuntu Core.

Detect if it's a snap only distro and in this case use snap:// URLs
instead of the appstream:// ones.