Add dynamic release announcement preview to update page The announcement title and description are parsed from the announcement page when the user consents to make the network request, or has opted-in to do so automatically. Automatic fetching is not done on a metered network, and can be retried in the event of a network failure.
For the kde.org site, we know how to get and insert the language code for it to try, and it will fallback to English if not available.

The beta/release update pages are now the same, both showing a preview of the announcement. As on the previous beta page, we show the spinning gear on the logo, have a different description and an additional bug report link. Instead of checking the patch version in QML, API has been added to Release for checking if this is a beta or development version to change behaviour or display.