Make pseudo-buttons based on AbstractButton
We have a number of UI elements that walk like a button and quack like a button,
i.e. they have some text, an icon, and trigger an action when clicked.

However the way they are implemented right now is not accessible, screen readers
don't recognize them as buttons and they don't participate in tab focus.

We can solve this by making them based on AbstractButton