Replace Systemsettings page with one about keyboard shortcuts + Meta key
The fact that you can customize things using System Settings is pretty
obvious, because every OS has that. So this page is not really pulling
its weight.

Instead, replace it with a more useful and Plasma-specific page about
handy systemwide keyboard shortcuts. In addition, we can take the
opportunity to explain to users what the "Meta" key is, because this
is not at all obvious if you don't already know.

There's also a mode where it shows a Mac-specific keyboard layout if it
detects you're using a Mac, since their keyboards have such a different
layout