pages/KDEConnect: Adapt to KCM -> App changes in KDE Connect
During Plasma 6.4's life, KDE Connect has changed such that the referenced KCM no longer exists. It now only provides the app.

In order for the "Open Settings…" action on the relevant page to work for users with up-to-date KDE Gear, rather than try to launch the KCM, we:
1. Try to launch the app (org.kde.kdeconnect.app)
2. Try to launch the KCM (kcm_kdeconnect, so we retain functionality for users with older KDE Gear)
3. If all else fails, show KDE Connect in Discover as it must not be installed (the action previously remained anyway, since fixed in 6.5)

This ensures the action works sensibly for both users with Gear 25.08 and users with earlier Gear.