Add logging categories and enhanced logging
This commit adds proper support for Qt logging categories to Welcome Center.

Additional logging is added to KDED for what we're doing when we're launching, as we've seen bugs or non-obvious behaviour in this area before.

In order to help with release preview localisation issues (see: CCBUG), logging is also added to Release::getPreview().

CCBUG: 511512